refactored minter into a module exporting methods to be used in multiple UIs
