+++ modules/cli/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@jfrazier-eth/cli",
+  "version": "1.0.0",
+  "type": "commonjs",
+  "engines": {
+    "node": ">=12"
+  },
+  "main": "./dist/index.js",
+  "exports": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc -b ./tsconfig.build.json",
+    "start": "yarn run build && node ./dist/index.js",
+    "start:dev": "yarn run build && cross-env MODE=development node ./dist/index.js"
+  },
+  "license": "GNU",
+  "dependencies": {
+    "@jfrazier-eth/minter": "^1.0.0",
+    "cross-env": "^7.0.3",
+    "dotenv": "^10.0.0",
+    "inquirer": "^8.2.0"
+  },
+  "devDependencies": {
+    "@types/inquirer": "^8.1.3",
+    "dotenv": "^10.0.0",
+    "typescript": "^4.5.2"
+  }
+}
