+++ modules/cli/src/prompts/promptAmountToMint.ts
