+++ modules/cli/src/prompts/promptMintMethod.ts
