+++ modules/cli/src/ui.ts
@@ -1,63 +1,20 @@
-import { ContractInterface, providers, BigNumber } from "ethers";
-import { Contract } from "../models/Contract.js";
-import { getContractABI } from "../services/etherscan.js";
-import { isDev } from "../utils/isDev.js";
-import { promptContractAddress } from "./promptContractAddress.js";
-import { promptJsonRpcUrl } from "./promptJsonRpcUrl.js";
-import { promptMintMethod } from "./promptMintMethod.js";
-import dotenv from 'dotenv';
-import { promptPricePerMint } from "./promptPricePerMint.js";
-import { setTerminalTitle } from '../utils/setTerminalTitle.js';
-import { promptAmountToMint } from "./promptAmountToMint.js";
+import {BigNumber, ContractInterface, providers} from "ethers";
+import {isDev} from "./utils/isDev";
+import {promptJsonRpcUrl} from "./prompts/promptJsonRpcUrl";
+import {Contract, getContractABI} from "@jfrazier-eth/minter";
+import {promptContractAddress} from "./prompts/promptContractAddress";
+import {promptMintMethod} from "./prompts/promptMintMethod";
+import {promptPricePerMint} from "./prompts/promptPricePerMint";
+import {promptAmountToMint} from "./prompts/promptAmountToMint";
 
-export async function setup() {
-    console.clear();
-
-    if(isDev()) {
-        dotenv.config();
-    }
-
-    const provider = await getProvider();
-
-    const contract = await getContract(provider);
-
-    const mintMethodName = await getMintMethod(contract)
-
-    const mintMethod = contract.interface.getFunction(mintMethodName);
-
-    if(!mintMethod.payable) {
-        throw new Error('selected method is not payable')
-    }
-
-    if(mintMethod.inputs.length !== 1) {
-        throw new Error("cannot mint from this contract");
-    }
-
-    const pricePerMintInWei = await getPricePerMint();
-
-    const amountToMint = await getAmountToMint();
-
-    const totalPrice = pricePerMintInWei.mul(amountToMint);
-
-    setTerminalTitle(`Amount to mint: ${amountToMint}. Total Price: ${totalPrice.div(BigNumber.from(10).pow(15)).toNumber() / 100} ETH`)
-
-    return {
-        provider,
-        contract,
-        mintMethodName,
-        pricePerMintInWei,
-        amountToMint,
-        totalPrice
-    }
-}
 
 /**
  * getProvider gets the user selected JSON RPC URL and returns
  * a provider
- * 
- * @returns a provider 
+ *
+ * @returns a provider
  */
-async function getProvider(): Promise<providers.JsonRpcProvider> {
+export async function getProvider(): Promise<providers.JsonRpcProvider> {
     let jsonRpcUrl;
 
     if(isDev()) {
@@ -66,7 +23,7 @@ async function getProvider(): Promise<providers.JsonRpcProvider> {
 
     if(!jsonRpcUrl) {
         jsonRpcUrl = await promptJsonRpcUrl();
-    } 
+    }
 
     const provider = new providers.JsonRpcProvider(jsonRpcUrl);
 
@@ -76,10 +33,10 @@ async function getProvider(): Promise<providers.JsonRpcProvider> {
 /**
  * getContract gets the user selected contract address, gets the contract abi,
  * and returns the corresponding contract
- * 
+ *
  * @param provider used to create the contract object
  */
-async function getContract(provider: providers.JsonRpcProvider): Promise<Contract> {
+export async function getContract(provider: providers.JsonRpcProvider): Promise<Contract> {
     let contractAddress;
 
     if(isDev()){
@@ -99,11 +56,11 @@ async function getContract(provider: providers.JsonRpcProvider): Promise<Contrac
 
 /**
  * getMintMethod gets the user selected name for the method to call to mint
- * 
+ *
  * @param contract containing methods to select from
  * @returns the selected method name
  */
-async function getMintMethod(contract: Contract) {
+export async function getMintMethod(contract: Contract) {
     let method;
     if(isDev()) {
         method = process.env.MINT_METHOD;
@@ -118,14 +75,14 @@ async function getMintMethod(contract: Contract) {
 
 /**
  * getPricePerMint gets the user entered price per mint in ether and converts it to wei
- * 
+ *
  * @returns the price per mint in wei
  */
-async function getPricePerMint(): Promise<BigNumber> {
+export async function getPricePerMint(): Promise<BigNumber> {
     let pricePerMintInETH;
     if(isDev()) {
         const price = parseFloat(process.env.PRICE_PER_MINT ?? "");
-        if(!Number.isNaN(price)){ 
+        if(!Number.isNaN(price)){
             pricePerMintInETH = price;
         }
     }
@@ -139,21 +96,19 @@ async function getPricePerMint(): Promise<BigNumber> {
     return priceInWei;
 }
 
-/**
- * 
- */
-async function getAmountToMint(){
+export async function getAmountToMint(){
     let amountToMint;
     if(isDev()) {
         const amount = parseInt(process.env.AMOUNT_TO_MINT ?? "", 10);
-        if(!Number.isNaN(amount)){ 
+        if(!Number.isNaN(amount)){
             amountToMint = amount;
         }
     }
-    
+
     if(amountToMint === undefined) {
         amountToMint = await promptAmountToMint();
     }
 
     return amountToMint;
 }
+
