+++ modules/cli/src/utils/isDev.ts
