+++ modules/cli/src/utils/setTerminalTitle.ts
