+++ modules/website/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "website",
+  "name": "@jfrazier-eth/website",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
