+++ modules/website/tsconfig.build.json
@@ -2,9 +2,11 @@
   "extends": "../../tsconfig.build.json",
   "compilerOptions": {
     "noEmit": true,
+    "isolatedModules": true,
     "jsx": "react-jsx"
   },
   "include": [
     "src"
-  ]
+  ],
+  "exclude": ["node_modules", "dist"]
 }
