+++ tsconfig.build.json
@@ -6,17 +6,15 @@
       "dom.iterable",
       "esnext"
     ],
-    "allowJs": false,
     "skipLibCheck": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
-    "module": "esnext",
+    "module": "commonjs",
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "isolatedModules": true,
   },
 
   "exclude": ["node_modules", "dist"]
